+++ app/middlewares/common.js
@@ -1,8 +1,6 @@
 const { checkSchema, validationResult } = require('express-validator');
 
-const { schemaError, notLoggedError } = require('../errors'),
-  { verifyJWT } = require('./utils'),
-  { TOKEN_START } = require('../constants');
+const { schemaError } = require('../errors');
 
 exports.validateResult = (req, res, next) => {
   const errors = validationResult(req);
@@ -13,21 +11,3 @@ exports.validateResult = (req, res, next) => {
 };
 
 exports.validateSchema = schema => [checkSchema(schema), exports.validateResult];
-
-exports.reportNotLoggedUser = (req, res, next, admin) =>
-  next(
-    notLoggedError(
-      admin ? 'You have to be logged in as an admin user' : "You don't have access, please login"
-    )
-  );
-
-exports.checkIfUserIsLogged = (req, res, next, admin) => {
-  let authorizationToken = req.headers.authorization;
-  if (authorizationToken && authorizationToken.startsWith(TOKEN_START)) {
-    authorizationToken = authorizationToken.substring(TOKEN_START.length);
-    return verifyJWT(authorizationToken, admin)
-      .then(() => next())
-      .catch(next);
-  }
-  return exports.reportNotLoggedUser(req, res, next, admin);
-};
