+++ app/middlewares/utils.js
@@ -2,7 +2,11 @@
 const { User } = require('../models'),
   { notLoggedError } = require('../errors'),
   logger = require('../logger'),
-  jwt = require('jsonwebtoken');
+  jwt = require('jsonwebtoken'),
+  { TOKEN_START } = require('../constants');
+
+const getNotLoggedError = admin =>
+  notLoggedError(admin ? 'You have to be logged in as an admin user' : "You don't have access, please login");
 
 exports.verifyJWT = (authorizationToken, admin) => {
   logger.info('Verifying JWT...');
@@ -18,24 +22,23 @@ exports.verifyJWT = (authorizationToken, admin) => {
         })
           .then(user => {
             logger.info('Search finished.');
-            return user
-              ? resolve()
-              : reject(
-                  // eslint-disable-next-line indent
-                  notLoggedError(
-                    admin
-                      ? 'You have to be logged in as an admin user'
-                      : "You don't have access, please login"
-                  )
-                );
+            return user ? resolve() : reject(getNotLoggedError(admin));
           })
           .catch(reject);
       }
-      return reject(
-        notLoggedError(
-          admin ? 'You have to be logged in as an admin user' : "You don't have access, please login"
-        )
-      );
+      return reject(getNotLoggedError(admin));
     });
   });
+};
+
+exports.checkIfUserIsLogged = (req, res, next, admin) => {
+  let authorizationToken = req.headers.authorization;
+  if (authorizationToken && authorizationToken.startsWith(TOKEN_START)) {
+    authorizationToken = authorizationToken.substring(TOKEN_START.length);
+    return exports
+      .verifyJWT(authorizationToken, admin)
+      .then(() => next())
+      .catch(next);
+  }
+  return next(getNotLoggedError(admin));
 };
